Version 2.0.0 Released
Notiflix "v2" is ready to use.

This version includes a new module name as "Block".

"Notiflix.Block.*" module can be used to block or unblock elements to prevents users actions during the process (AJAX etc.) without locking the browser or the other elements.

Added: The "Block" module added to Notiflix.

Added: The "clickToClose" option added to the "Notiflix.Notify.*" module. Default is "false". This option can be set as "true" to remove each notification without waiting for it to be automatically removed when it has been clicked.

Changed: Code Review.